pyproject.toml
[project]
name = "mywebsite-load-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "locust>=2.20.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
]
